Rewrite Tabbar
